feat: unify header proof fetching + group header proofs by mmr meta

+++ crates/types/src/keys/starknet/storage.rs
@@ -52,7 +52,7 @@ pub type StorageKey = Felt252;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
 pub struct Key {
-    pub chain_id: u128,
+    pub chain_id: u64,
     pub block_number: u64,
     pub address: Felt252,
     pub storage_slot: Felt252,
